+++ lib/hexapdf/cli.rb
@@ -46,6 +46,7 @@ require 'hexapdf/cli/batch'
 require 'hexapdf/cli/split'
 require 'hexapdf/cli/watermark'
 require 'hexapdf/cli/image2pdf'
+require 'hexapdf/cli/form'
 require 'hexapdf/version'
 require 'hexapdf/document'
 
@@ -99,6 +100,7 @@ module HexaPDF
         add_command(HexaPDF::CLI::Split.new)
         add_command(HexaPDF::CLI::Watermark.new)
         add_command(HexaPDF::CLI::Image2PDF.new)
+        add_command(HexaPDF::CLI::Form.new)
         add_command(CmdParse::HelpCommand.new)
         version_command = CmdParse::VersionCommand.new(add_switches: false)
         add_command(version_command)